timeout query param is only available from v7.17.0
