+++ internal/clients/elasticsearch/index.go
@@ -544,169 +544,3 @@ func DeleteIngestPipeline(ctx context.Context, apiClient *clients.ApiClient, nam
 	}
 	return diags
 }
-
-func PutTransform(ctx context.Context, apiClient *clients.ApiClient, transform *models.Transform, params *models.PutTransformParams) diag.Diagnostics {
-
-	var diags diag.Diagnostics
-	transformBytes, err := json.Marshal(transform)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	esClient, err := apiClient.GetESClient()
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	putOptions := []func(*esapi.TransformPutTransformRequest){
-		esClient.TransformPutTransform.WithContext(ctx),
-		esClient.TransformPutTransform.WithDeferValidation(params.DeferValidation),
-		esClient.TransformPutTransform.WithTimeout(params.Timeout),
-	}
-
-	res, err := esClient.TransformPutTransform(bytes.NewReader(transformBytes), transform.Name, putOptions...)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	defer res.Body.Close()
-	if diags := utils.CheckError(res, fmt.Sprintf("Unable to create transform: %s", transform.Name)); diags.HasError() {
-		return diags
-	}
-
-	if params.Enabled {
-		startOptions := []func(*esapi.TransformStartTransformRequest){
-			esClient.TransformStartTransform.WithContext(ctx),
-			esClient.TransformStartTransform.WithTimeout(params.Timeout),
-		}
-		startRes, err := esClient.TransformStartTransform(transform.Name, startOptions...)
-		if err != nil {
-			return diag.FromErr(err)
-		}
-
-		defer startRes.Body.Close()
-		if diags := utils.CheckError(startRes, fmt.Sprintf("Unable to start transform: %s", transform.Name)); diags.HasError() {
-			return diags
-		}
-	}
-
-	return diags
-}
-
-func GetTransform(ctx context.Context, apiClient *clients.ApiClient, name *string) (*models.Transform, diag.Diagnostics) {
-
-	var diags diag.Diagnostics
-	esClient, err := apiClient.GetESClient()
-	if err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	req := esClient.TransformGetTransform.WithTransformID(*name)
-	res, err := esClient.TransformGetTransform(req, esClient.TransformGetTransform.WithContext(ctx))
-	if err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	defer res.Body.Close()
-	if res.StatusCode == http.StatusNotFound {
-		return nil, nil
-	}
-	if diags := utils.CheckError(res, fmt.Sprintf("Unable to get requested transform: %s", *name)); diags.HasError() {
-		return nil, diags
-	}
-
-	transformsResponse := models.GetTransformResponse{}
-	if err := json.NewDecoder(res.Body).Decode(&transformsResponse); err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	for _, t := range transformsResponse.Transforms {
-		if t.Id == *name {
-			t.Name = *name
-			return &t, diags
-		}
-	}
-
-	return nil, diags
-}
-
-func UpdateTransform(ctx context.Context, apiClient *clients.ApiClient, transform *models.Transform, params *models.UpdateTransformParams) diag.Diagnostics {
-
-	var diags diag.Diagnostics
-	transformBytes, err := json.Marshal(transform)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	esClient, err := apiClient.GetESClient()
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	updateOptions := []func(*esapi.TransformUpdateTransformRequest){
-		esClient.TransformUpdateTransform.WithContext(ctx),
-		esClient.TransformUpdateTransform.WithDeferValidation(params.DeferValidation),
-		esClient.TransformUpdateTransform.WithTimeout(params.Timeout),
-	}
-
-	res, err := esClient.TransformUpdateTransform(bytes.NewReader(transformBytes), transform.Name, updateOptions...)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	defer res.Body.Close()
-	if diags := utils.CheckError(res, fmt.Sprintf("Unable to update transform: %s", transform.Name)); diags.HasError() {
-		return diags
-	}
-
-	if params.Enabled {
-		startOptions := []func(*esapi.TransformStartTransformRequest){
-			esClient.TransformStartTransform.WithContext(ctx),
-			esClient.TransformStartTransform.WithTimeout(params.Timeout),
-		}
-		startRes, err := esClient.TransformStartTransform(transform.Name, startOptions...)
-		if err != nil {
-			return diag.FromErr(err)
-		}
-
-		defer startRes.Body.Close()
-		if diags := utils.CheckError(startRes, fmt.Sprintf("Unable to start transform: %s", transform.Name)); diags.HasError() {
-			return diags
-		}
-	} else {
-		stopOptions := []func(*esapi.TransformStopTransformRequest){
-			esClient.TransformStopTransform.WithContext(ctx),
-			esClient.TransformStopTransform.WithTimeout(params.Timeout),
-		}
-		stopRes, err := esClient.TransformStopTransform(transform.Name, stopOptions...)
-		if err != nil {
-			return diag.FromErr(err)
-		}
-		defer stopRes.Body.Close()
-		if diags := utils.CheckError(stopRes, fmt.Sprintf("Unable to stop transform: %s", transform.Name)); diags.HasError() {
-			return diags
-		}
-	}
-
-	return diags
-}
-
-func DeleteTransform(ctx context.Context, apiClient *clients.ApiClient, name *string) diag.Diagnostics {
-
-	var diags diag.Diagnostics
-	esClient, err := apiClient.GetESClient()
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	res, err := esClient.TransformDeleteTransform(*name, esClient.TransformDeleteTransform.WithForce(true), esClient.TransformDeleteTransform.WithContext(ctx))
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	defer res.Body.Close()
-	if diags := utils.CheckError(res, fmt.Sprintf("Unable to delete transform: %s", *name)); diags.HasError() {
-		return diags
-	}
-
-	return diags
-}
